:truck: Combine Solution Explorer Panel And Tab
