--- src/modules/solution-explorer/solution-explorer-tab/solution-explorer-tab.scss
@@ -1,37 +1,0 @@
-.solution-explorer-tab {
-  position: relative;
-  height: 100%;
-}
-
-.solution-explorer-tab__header {
-  padding: 5px 5px;
-  display: flex;
-  color: #a5a5a5;
-  border-bottom: 2px solid #dcdbdb;
-}
-
-.solution-explorer-tab__header__name {
-  font-size: 16px;
-  flex: 1;
-}
-
-.solution-explorer-tab__actions {
-  display: flex;
-}
-
-.solution-explorer-tab__actions button {
-  padding: 0 2px;
-  background: transparent;
-  border: none;
-  cursor: pointer;
-  vertical-align: middle;
-}
-
-.solution-explorer-tab__actions button:hover {
-  color: #5e5e5e;
-}
-
-.solution-explorer-tab__input {
-  visibility: hidden;
-  height: 0;
-}
